+++ DropboxSDK/MPURLRequestParameter.m
@@ -80,7 +80,7 @@
 + (NSString *)parameterStringForParameters:(NSArray *)inParameters {
 	NSMutableString *queryString = [[NSMutableString alloc] init];
 	int i = 0;
-	int parameterCount = [inParameters count];	
+	NSUInteger parameterCount = [inParameters count];
 	MPURLRequestParameter *aParameter = nil;
 	
 	for (; i < parameterCount; i++) {
